Fix docker config handling in offline installer
This fixes adding of dns address to
/etc/docker/daemon.json to always use infra's ip address
instead of possibly using it's hostname.
It is done by replacing usage of "ansible_host" by trying to guess ip address to use.
It will prefer ip address of interface to which default route points to.
Othervise it will use first ip of first interface.
Later some variable can be added to override that.

Change-Id: I2baa3f4c7d69b8dff109b5f8ae984134caf2af7d
Issue-ID: OOM-1573
Signed-off-by: Michal Zegan <m.zegan@samsung.com>