Fix docker config handling in offline installer
This fixes adding of dns address to
/etc/docker/daemon.json to always use infra's ip address
instead of possibly using it's hostname.
Also similar change is done in handling of
/etc/resolv.conf in infrastructure.yml.
In general, the fix is to specify a subnet where
kubernetes nodes will communicate, and use the infra's ip address that matches
this subnet.
The list of ip addresses is taken from ansible_all_ipv4_addresses fact.

Change-Id: I2baa3f4c7d69b8dff109b5f8ae984134caf2af7d
Issue-ID: OOM-1573
Signed-off-by: Michal Zegan <m.zegan@samsung.com>