Update patch set 4
Patch Set 4: -Code-Review

It seems like we took the longer route to applicate my proposal and still we are not there yet. I fail to see what using subnet instead of ip brings - maybe it is the correct approach what solves some other issue but I am missing explanation. I see it again as a possible place for problem - you are filtering ips and what if user have multiple ips on that subnet and it filters out the ip he actually wants? I just really dont understand why we are not using the obvious option and go by the fishy route. I am removing my vote because this patchset basically does what I proposed but I am still unhappy that we are denying the user to pick exactly the ip he wants...

Patch-set: 4