Update patch set 4
Patch Set 4:

I can, I did not want to hijack your patchset - on the topic - there is a very common usage when you do use multiple ips on one interface and in one subnet - the floating ip - user can choose to have infra node as HA - that means on that same subnet he can have basic ip for the infra-node1 (not-floating) and another floating ip on the same subnet which is added/deleted when failover is done and which is migrating between two or more nodes (pacemaker, keepalived etc.). In this case the floating ip would be the one ansible should have configured - but probably it would not be selected due to this autoconfigure. I just dont understand why we are making this flaw here and what is the benefit to have subnet instead of explicit ip. It is not that I want to have this change exactly as I proposed if this solution is superior but I dont see any upside and I already point out usecase when it has big drawback.

Patch-set: 4