Fix the problem with the firewall
As of now the installer does not configure iptables rules and for
successfull installation some rules are needed. This is simple fix
to enable installation even on machines with enabled firewall.

The user must setup rules by himself for a production use.

Firewall setup may be added in a future revision of the installer.

Change-Id: I7ea741b088dbb7f64bb7e8e273f7648f7fbf1bd6
Issue-ID: OOM-1581
Signed-off-by: Petr Ospalý <p.ospaly@partner.samsung.com>