Enable to build ansible image out of git directory
Inside the build_ansible_image.sh is a variable (label for docker image)
which is set to git commit hash. This is fine but you can just copy
and move ansible directory elsewhere out of git and this script breaks
because of this.

This change preserve previous functionality but only if the script
is run inside a git repository otherwise it will skip setting the
label for the built docker image.

Change-Id: I23f828542efddde21d1c96116de7b20272ef23ef
Issue-ID: OOM-1575
Signed-off-by: Petr Ospalý <p.ospaly@partner.samsung.com>