Update patch set 1
Patch Set 1:

This: git checkout N/A

What is purpose of this command? If this directory is outside of git then echo N/A is not run at all - if it is in git directory then the original git rev-parse is used and because it will be valid it can be safely used. The output is used as a label - no one is interpreting that label if it is outside of git. So I dont see any relevancy with your artificial command - can you show me usage? That is like artificial example of erasing a random line from script and show it as a proof that the script is fundamentally flawed.

I see this as a very simple change and the ratio of discussion vs the actual change is really bizarre :)

If I abondon this change then your artificial example will turn to this:

$ git checkout

See missing argument and also the script stops working.

Pick your poison.

Patch-set: 1
Label: Code-Review=0