Update patch set 1
Patch Set 1:

@Samuli Silvius
That's exactly what I thought about while writing first comment.
Let's add git-commit label only if we can generate it.
Not to mention that it would be good to call the git command from script_dir as currently it works only if we call a script from inside of git repo. (And BTW if we call it from another git repo we will get hash from that repo not from offline installer one;)

Patch-set: 1