Update patch set 5
Patch Set 5:

If the future of offline-installer is to really support various linux distros then creating a common config file might be the least of our worries. I'd say we should refactor the build process to be more generic to allow support for different platforms and looking even further i think we could use ansible for that. There's plenty of bashism so no go for moving all of the build logic to ansible, but at least we could wrap the main flow of the build process in ansible and call scripts from there.

Patch-set: 5