Update patch set 1
Patch Set 1:

> I also find it now less readable - like I said before - I need a
 > "find" and "grep" to understand ansible. I am not sure what is the
 > ansible way, but so far it is quite scary. I understand that
 > over-use of the shell module is a bad practice (my bad), but using
 > non-declarative, procedure-like constructs (loop) seems like
 > writing a procedural code in a very ugly and not much expressive
 > language - ansible. But I am still ansible novice so I feel
 > incompetent to say what is a good or bad ansible.

It should not be less readable now as things are defined and used now only where they are needed. Maybe if I change variable passing to global as suggested by Bartek you will find it better? Btw what IDE you are using. At least using purely vi for opening one file at the file is not a best option.

Last part of you comment is not targeted to this commit. Yes ansible is not a programming language, it's declarative configuration language. And writing e.g. loops as you mentioned is not right for it's nature.
However it's still better than the previous code where all instances of variables was explicitly written and same code was copy-pasted for the different variable.
Any mode difficult functionality should be written with own ansible module (where real programming language can be used) and in ansible we are back in declarative way. But in this role we are not even close to that complexity.

We still have that shell call there for the actual ssh transfer. And that would be a worth of writing own ansible module for that if existing ansible module cannot handle the job. Because of that one shell task idempotence is handled here quite simple way and already on the main module of the role (just check if dir is already there) and that's not very good. With own module idempotence could be handled more correctly.

Patch-set: 1
Label: Code-Review=0