Update patch set 2
Patch Set 2:

@samuli - I think my problem is with the fact that ansible is trying to be declarative (using dummy configuration file format - yaml), but fails there because you still end up telling it HOW to do stuff instead of WHAT it should do.

Like I want to fix a problem with this playbook, which fails when the ssh-key is protected with a passphrase.

Ideal declarative way:

Host: infra
role: extract_resources(from_host, to_dir)

Ansible shoudl figure out how to achieve that (or that role/module).

But in reality to achieve that I need several steps and you cannot even write it as a module because the steps are spread over two hosts (or can you?):

1. create ephemeral ssh-keypair on infra
2. add a public key of this just created key to authorized_keys on resource-host
3. do the copying and extraction on infra - that is implemented already
4. delete this private key on infra
5. remove it from authorized keys on resource-host

Can you hide all of this this in one module/role?

If so, then ok. But I think you cannot.

So at the end, what ansible brings to the table? You could just use a python program from get go and utilize modules - which at the end is what ansible does...but it is obscured first with a bunch of yamls, where they invented horrible syntax to do loops etc.

Btw. I can compare ansible to saltstack which I used before and there is very much often used jinja templating - and it feels much better when you need to do some imperative logic.

Patch-set: 2