Update patch set 1
Patch Set 1: Code-Review-1

(2 comments)

I don't really see the value of this particular playbook to be moved to role at this poomt. But it is possible that we will want add something more complicated in the future (doubt it) so I am willing to let it be in role.
But please fix hardcoded key name.

Patch-set: 1
Label: Code-Review=-1