Update patch set 3
Patch Set 3:

@mateusz: it will have effect on newly created containers, but the already created containers (before setting daemon.jso and docker restart) are not affected - you can check this with docker inspect <container>

Also it is possible that rancher/helm is effectively ignoring this settings when they deploy containers - it must be tested for the whole installation procedure. There have been a plenty of new merges in the master so some testing installation will be fine - I may ge to it, but not so soon.

Patch-set: 3