Update patch set 3
Patch Set 3:

About "Also it is possible that rancher/helm is effectively ignoring this settings when they deploy containers - it must be tested for the whole installation procedure. There have been a plenty of new merges in the master so some testing installation will be fine - I may ge to it, but not so soon."

For us it is the most important that kubernetes pods (kube-system) are using log rotation/limits as it was one of them that used all storage.

Patch-set: 3