Update patch set 1
Patch Set 1:

> > Patch Set 1:
 > >
 > > well there will be. So you need a reminder to do reverse commit
 > when we are ready for casablanca and will create casablanca branch.
 > 
 > But i thought that for package build we should always use master,
 > don't we? BuildGuide is general, not Onap release specific i think?

If user is reading casablanca version of this guide (from casabanca branch which does not exist yet) then cloning master is not correct.
But I'm favor of removing this kind of items that needs to be always updated in each release. So the cloning step can be generalized here and rely that user is wise enough to clone correct branch.

Patch-set: 1
Reviewer: Gerrit User 3957 <3957@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Verified=0