Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> this is not working as expected, did you try to break repo ? This update-cache is always passing even when I moved rpms away
> This system is not registered with an entitlement server. You can use subscription-manager to register.
> file:///mnt/opt/onap/pkg/rhel/repodata/repomd.xml: [Errno 14] curl#37 - "Couldn't open file /mnt/opt/onap/pkg/rhel/repodata/repomd.xml"
> 
> I would recommend some real check even w/o yum module like another check
> like "command: "yum repo-pkgs {{ item.name }} list""  from
> https://gerrit.onap.org/r/#/c/81834/1/ansible/roles/package-repository/tasks/main.yml

Broken repo it the context of this task/patch means "broken or missing repo metadata (eg repomd.xml), not RPMS themself!" and that was the scope of the task and aim of this patch.

In comment above i pasted a screen dump (in OOM-1689) of the ansible behaviour in case of missing metadata. Checking for existence of particular RPMS in this task is undesired because then we would end up with one more place with list of packages to add new packages to (packaging script, actual task it is installed in and this task with  checks). 

This task is to detect if repo itself is working and fail fast in case it's not. If RPMs are missing then particular task where they are installed will fail and that's expected.

And secondly: Refreshing package cache is completely distinct from REBUILDING REPO itself.

Patch-set: 1