Rewrite the run_chroot.sh script
This change heavily changes run_chroot.sh script.
Major change is that most of the real script logic is now run in a namespace,
so that all mounts are done in namespace.
Also, all the cleanup code and some checks are removed, also the
shell wrapper is removed.
This change is for simplifying the run_chroot.sh. It will function the same as before
from the user's point of view, except that now the script does not have to explicitly
unmount filesystems that it mounted before, so
no cleanup code isneeded. Also,
all processes run from inside a chroot should be killed immediately
when the chroot command finishes.

Change-Id: I45a3dc7c6583cb9693746cc04f42b7af325724dd
Signed-off-by: Michal Zegan <m.zegan@samsung.com>
Issue-ID: OOM-1744