Update patch set 1
Patch Set 1:

Well, I take a completely different approach here, instead of having an override role that may have any logic it desires, i just assume that all overrides can be statically passed by -e or application_configuration.yml or similar, and the only exception is the root certificate that needs to be merged with this.
Not quite sure what is the use case for a separate override role then.

Patch-set: 1