Update patch set 1
Patch Set 1:

> Well, I take a completely different approach here, instead of
 > having an override role that may have any logic it desires, i just
 > assume that all overrides can be statically passed by -e or
 > application_configuration.yml or similar, and the only exception is
 > the root certificate that needs to be merged with this.
 > Not quite sure what is the use case for a separate override role
 > then.

Yes but I think those approaches can be merged. Somehow we should merge anyway.

First use case I wanted to add is the ability to skip override and not to have it mandatory.

Second use case I added is to move it to own role meaning user can write own role and code to generate the override file. That means override don't need to be static.

Third use case I added is the possibility to define path and name for the override file.

Patch-set: 1