Update patch set 1
Patch Set 1:

Generic question/doubt: Centos7 is our secondary platform, RedHat being the primary one. Shouldn't we have dockerfile for the latter one in the first place? I mean i'm aware they are sibling platforms but various differences still apply though.

Or maybe it is the case that we develop using Centos image and rely on ci/cd with redhat image to be a final verification step before merge? Because if we introduce molecule testing on ONAP's ci/cd then i guess we'll be using this Centos image, not RedHat? What about RedHat OAT then?

Patch-set: 1