Update patch set 1
Patch Set 1:

(3 comments)

> Generic question/doubt: Centos7 is our secondary platform, RedHat
 > being the primary one. Shouldn't we have dockerfile for the latter
 > one in the first place? I mean i'm aware they are sibling platforms
 > but various differences still apply though.
 > 
 > Or maybe it is the case that we develop using Centos image and rely
 > on ci/cd with redhat image to be a final verification step before
 > merge? Because if we introduce molecule testing on ONAP's ci/cd
 > then i guess we'll be using this Centos image, not RedHat? What
 > about RedHat OAT then?

Yes very correct points, thanks.

This centos7 Dockerfile is just grapped from internet the first one I found with ssh and systemd pre-installed (reference in the Dockerfile first line) and I think I have not modified it at all.
So it's just to get testing ongoing and team (you all) into it to make it better or even better write it all from scratch to exactly meet our purposes and a close as possible to real VM image (for centos6/7 we are going to support).

Yes RedHat would be the first one we need. But same as with Centos, I did not had time or skills to write it and especially here would like the team/community to help.

Same with Ubuntu Dockerfile.

Patch-set: 1