Update patch set 1
Patch Set 1:

Ok i didn't know it's just a draft taken as is from somewhere else. So if that's the case then we can improve it collaboratively, sure.

But since we want a container that will resemble a target platform as much as possible, is the Docker the right tool of our choice? Isn't Vagrant what we would really need (with Vbox/Qemu provider)? Or on the other hand in Vagrant the provisioning time would be too long and no way to have fast rollbacks (like with Docker images), correct?

Patch-set: 1