Update patch set 1
Patch Set 1:

> Ok i didn't know it's just a draft taken as is from somewhere else.
 > So if that's the case then we can improve it collaboratively, sure.
 > 
 > But since we want a container that will resemble a target platform
 > as much as possible, is the Docker the right tool of our choice?
 > Isn't Vagrant what we would really need (with Vbox/Qemu provider)?
 > Or on the other hand in Vagrant the provisioning time would be too
 > long and no way to have fast rollbacks (like with Docker images),
 > correct?

I think docker is the right one for the testing we intend to do. Like you already answered Vagrant would be just so much heavier that the motivation for developer to develop/run/test on own laptop instead of reserving "real lab" (expensive and limited in amounts resource) is too low.
With docker driver we can test the functionality very well and leave rest of the integration to real labs (nighly).
Also Vagrant driver would require laptop's main operating system to be Linux which usually is not the case but developer runs linux on Virtualbox. Or run it in Windows laptop that surely should work but not so attractive place for our ansible project.

3rd reason is that Vagrant driver is much more in beta level in molecule compared to docker driver.

But outside of developer testing we surely could setup additional testing pipes in addition to nighly to run same tests with Vagrant driver on CI system at some point.

Patch-set: 1
Label: Code-Review=0