Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> @Michal: Good comments, like also Milan commented that instead of template we should use lineinfile module not to destroy existing entries.
> Just that none of the comments relate to this commit, but the original code itself:)
> I can improve those however as part of this molecule commit or in separate commit?

Yes, my personal opinion on that could be that code reviewing should be commenting on changes made, not improvements. I mean they can be pinpointed but should be the subject to separate patch/task/story, not blocker for current patch in review. Otherwise we're distorting the idea of code reviewing IMO.

Patch-set: 3