Update patch set 3
Patch Set 3:

in fact the comment about cluster_ip is not directly about this commit, but related because this commit modifies the logic to take a host from ansible_host to have a default, and the comment is related to this same logic. not sure what about such case.

Patch-set: 3