Update patch set 1
Patch Set 1:

> name of the file is the tag name...
yes it is. I don't think we need to change it on that level. The original idea was to offer 3.0.0 version of original casablanca and 3.0.1 version of casablanca with service pack ... due to wrong handling of those tags we better have there latest casablanca code

Patch-set: 1
Label: Code-Review=0