Update patch set 9
Patch Set 9:

Yes, I'd like to separate the rancher in other commit, since here's a lot of changes already. I don't get it with the naming. Since all variables I saw within ansible dir was with the underscore. So we should use dash in case of container/image names?

Patch-set: 9