Update patch set 1
Patch Set 1: Code-Review+1

About the tag: this is my proposal for the tagging scheme, it's an official way of tagging Centos images: name=centos, tag=<release> eg 'centos7'.

Then in molecule.yml we would have:
  image: molecule-${PREBUILD_PLATFORM_DISTRO:-centos}:${PREBUILD_DISTRO_VERSION:-centos7.6}


As for the Dockerfile - verified it with some already created uploads with molecule tests and it works even better that previous one.

Patch-set: 1
Label: Code-Review=+1