Update patch set 1
Patch Set 1:

All changes in this chain have been tested (successfully to the point that kubernetes is running, kubectl works and dashboard works). This list of docker images should be added to a proper place - I loaded them manually to the nexus on the infrastructure server only to be sure that nothing else was pulled from the internet. So far so good.

Patch-set: 1
Label: Code-Review=0