Update patch set 2
Patch Set 2:

Would be happy to add something like kubernetes cluster healthcheck, after this change.
Reasoning is that rke actually does ectd/kubelet/whatever healthchecks, but does not healthcheck pods in kube-system. Not sure how important it is for us to check that, the main source of error there is missing docker images. like dashboard or whatever images.

Patch-set: 2