Update patch set 2
Patch Set 2:

> > This moves kubeconfig handling from kubectl role to rancher role.
 > > Of course that is because this handling differs significantly, so
 > > the respective rancher/rke roles should create kubeconfig.
 > Although
 > > it seems a bit weird to me that kubectl role installs kubectl and
 > > does nothing else now. ... well same as helm role, so maybe I
 > > should not care.
 > 
 > I don't understand this. How it differs? From this commit does not
 > seem to differ at all.
 > Could we use kuctl role from rancher and rke roles and not to
 > copy-paste implementation?

well rke handling of kubeconfig is completely incompatible with rancher's handling of kubeconfig.
So the current handling was removed from kubectl role and put into the rancher role, and rke role has it's own.
Specifically, rancher role generates kubeconfig, but the rke role does not, it is generated by rke itself when it creates a cluster, and rke role only copies it to target.

Patch-set: 2
Label: Code-Review=0