Add docker preparation role (dind version)
This change adds the prepare-docker-dind role intended to
be used in tests to install docker on test containers.
Contrary to the prepare-docker role it arranges for docker daemon to be launched inside
of the test containers instead of using/polluting a test host.
It is a prerequisite to the rke role molecule tests.

Change-Id: I6be270758b01f7e9c4fa09aebcc32930c0977ea3
Issue-ID: OOM-1778
Signed-off-by: Michal Zegan <m.zegan@samsung.com>