Update patch set 7
Patch Set 7:

Hi, this change currently has a major issue that I didn't think about. Not quite sure how best to solve it in this case.
Namely: cluster.yml.j2 template will put control plane on the first kubernetes node in the kubernetes group.
However, it seems that we cannot predict which node it is. That is because although lists (like groups['kubernetes'] are ordered, dicts are not, and contents of the groups['kubernetes'] come from the inventory that is actually a dict, and dicts are not ordered.
I hit this during molecule testing, as the node2 was put before node1 in the group.
That is not only bad because it is unpredictable and does not allow doing tests on the control plane node, but you also can't predict where is your api server. If you want to access the api server externally using some form of external ip like floating ip in openstack, you are essentially not able to do that because you don't know in advance which node is your api server, it depends on python impl details.
We could either make a special group specifying control plane, or move control plane (and only it) to the infra server.
What do you think?

Patch-set: 7