Update patch set 7
Patch Set 7:

I found it funny because this is basically similar to the problem with referencing index zero (no one knows what machine we are actually making special :)

I made the control plane to be on the exactly one node due to the problem with not able to have control plane together with etcd in multinode cluster (mentioned in the commit message). There are two solution: put a name on it (which everyone "argued" against it...) or we will adjust the inventory to the specifics to the rke - it is not vanilla kubernetes anyway, it is again rancher special implementation with all their lingo and technical terms - so, we can create inventory where we differentiate singlo-node vs multi-node cluster and where we designate control plane nodes, etcd backends and workers...that is what I propose.

Patch-set: 7
Label: Code-Review=0