Fix idempotence issue in dns role
Docker_container module was called twice: in a taks
'Start dns server container' and then as a handler
but there was a divergent set o parameters for the
module in them. This had the efect of the container
being started by the means of the task itself but
immediately after it was killed and started again
by the handler because it had different set of
arguments for the module. The fix is to include the
handler itself as a task so the whole run is idempotent.

Change-Id: I9078f9dfe5894d94d4b6a9e44411f034508992e7
Issue-ID: OOM-1845
Signed-off-by: Bartek Grzybowski <b.grzybowski@partner.samsung.com>