Update patch set 9
Patch Set 9:

> > The error was: 'ansible.vars.hostvars.HostVarsVars object' has no
 > > attribute 'ansible_ssh_host_key_rsa_public'
 > 
 > interesting, works for me. any reproducer?

we found that if this playbook is executed separately, facts from k8s nodes are not collected but required by first play in rke.yml playbook. This is not acceptable and easily reproducable. The reason why it works for MichalZ is probably that he is always running site.yml.

Patch-set: 9
Label: Verified=0