Update patch set 2
Patch Set 2:

well no. this role is meant to replace prepare-docker in the long run, or it is at least a proposal/alternative implementation without destroying prepare-docker and other tests using it. I believe that either prepare-docker and this role should be merged, or if there are any objections on how this role works and good reasoning for how prepare-docker works, rke role should use prepare-docker instead. In case it is unclean, both could coexist for some time. But it is definitely not meant to be rke specific.

Patch-set: 2