Update patch set 2
Patch Set 2:

> I haven't seen any task on Jira with a description/proposal of an
 > initiative to rework prepare-docker role in a way that all tests
 > will utilize DinD. Without a task/epic for that it won't happen
 > unless you're driver for that?
 > 
 > So if that is the case please raise appropriate tasks on Jira,
 > otherwise it won't happen unless you're on the track with
 > refactoring all tests to use dind which i haven't heard of.
 > 
 > "prepare-docker" role has much more functionality than what's in
 > this commit.

Well, I just quickly made a separate role because it was nicer for me than writing cleanup playbook that would require careful collecting of image names and such like. however you are right that making a dind version of prepare-docker as the only option would require modifications to all other roles and would varrant a new task. I can be a driver for that, however not sure what should be the next steps in this specific case.

Patch-set: 2
Label: Code-Review=0