Update patch set 10
Patch Set 10:

This patchset contains one thing that we didn't discuss.
Previously rke user was created on infra only because it *may* be useful when moving k8s components like controlplane to infra, and othervise this user was only used as an excuse to generate ssh keys that would be distributed to nodes.
However because it was decided to have a completely separate control plane group (here called kubernetes-control-plane), this argument became invalid.
For that reason what this change introduces is not to create any user on infra, but generate ssh key using ssh-keygen (the key is stored along with other cluster configs in /opt/onap/cluster).
Then, the rke_node.yml tasks are not only executed on worker nodes, but also on control plane nodes.
That way, if infra is set as control plane (or worker), it is processed like a node, the user is created and ssh key is authorized. If infra is not mentioned, this does not happen. I believe it makes far more sense.
By the way, I believe that the way it is done, you may currently be able to colocate infra, controlplane and worker on the same instance. You may also separate control plane from both infra and workers. The only thing you may not do (but it is now very easy to add) is to separate etcd's from workers.

Patch-set: 10