Update patch set 6
Patch Set 6:

> why destroy.yml is needed? Could we contribute to molecule? What is
 > missing?

It is needed because in case of this specific role, the containers have anonymous volumes created with them. The default destroy.yml of molecule does not take them down along with the container.
This version of destroy.yml is a verbatim copy of molecule's own, just with added flag to remove volumes during container removal. Othervise we waste really a lot of space.

Patch-set: 6