Update patch set 8
Patch Set 8:

explanation: note that I moved it from kubectl role to rancher_kubernetes for a reason. rancher_server.yml does *create* the kubeconfig by generating it from template. It is in contrast to the one in rke_deploy.yml that copies an existing kubeconfig file that was generated by rke directly. So the only thing that is identical may be creating ~/.kube, but probably it is not enough to varrant a separate task file, is it?

Patch-set: 8
Reviewer: Gerrit User 5019 <5019@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0