Update patch set 12
Patch Set 12:

> Patch Set 11:
> 
> After rebasing on all recent changes it still fails on
> 
>     fatal: [infrastructure-server]: FAILED! => {"attempts": 30, "changed": false, "msg": "Failed to connect to nexus.infrastructure-server at port 443: [Errno 111] Connection refused"}
>     
>     PLAY RECAP *********************************************************************
>     infrastructure-server      : ok=70   changed=38   unreachable=0    failed=1
> 
> Samuli did you do any investigation on that? AFAIR even previously we encountered that issue with connecting to nexus.

I just tried, but it failed for me to same. I don't know what broke it, it was working for me previously. I don't have time now to back to deep investigation, so disabled now that images pushing (it's not tested by default in real env either e.g. nightly). So somebody could even manually test it in real env, maybe it's broken (but more likely related to molecule) as nightly does not test it.

Patch-set: 12
Label: Code-Review=0