prepare-docker-dind: make docker start optional
This change makes it possible to disable docker start after
prepare-docker-dind role executes.
Needed for docker role that wants to start docker by itself.
Note that the docker role also tries to install docker by itself, and
prepare-docker-dind does the same.
The reason is that it makes it possible to fix the version of installed docker.

Change-Id: I7487bc46b316c0e6c782b22d8e24afea35c3c2d5
Issue-ID: OOM-1863
Signed-off-by: Michal Zegan <m.zegan@samsung.com>