Update patch set 15
Patch Set 15:

(2 comments)

> Patch Set 15:
> 
> (5 comments)
> 
> Why docker images are still downloaded with host docker engine in prepare- roles? If dind can it be also done on infra docker?
No. Part of infra playbook is installation of docker, so you don't expect docker to be there at the time the playbook is imported. And especially you do not expect it to be started. That is why I call host docker, because we have a guarantee to have docker in the host *already*. The only requirement is for the user running ansible to be in docker group, but that seems to be met even for the molecule-dev docker thing, so no problem.
> 
> Should prepare-play-infrastructure role functionality be in prepare-nexus role instead?
Ned to check, likely yes.

Patch-set: 15
Label: Code-Review=0