Update patch set 3
Patch Set 3:

(2 comments)

Super, already wanted to change the way of configuration this way and I like it, just wanted to have possibility to use more lists (similar as Milan did now). Not sure what we gonna do with the tag. We can remove that in next commit once there will be a consensus

Patch-set: 3
Label: Verified=0