Allow unsafe writes when modifying resolv.conf
The package-repository role edits /etc/resolv.conf.
Ansible modifies files atomically to prevent data loss on crash, but that
does not work in docker containers because /etc/resolv.conf is
bind-mounted.
Allow ansible to fall back to non atomic writes in such cases.
This makes it unnecessary to hack around this.

Change-Id: I066205322de5cc06569e6ea3aa38885d3834f9cd
Issue-ID: OOM-1816
Signed-off-by: Michal Zegan <m.zegan@samsung.com>