Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> If you are changing the file name here, perhaps you should consider also editing build/build_nexus_blob.sh line 106: NXS_DOCKER_IMG_LIST="${LISTS_DIR}/${TAG}-docker_images.list".

This is not the complete list yet. To get the full list you need to append rke list to this. It can be update in the build guide. Or modify the build/build_nexus_blob.sh to use more lists to populate the nexus

Patch-set: 1
Reviewer: Gerrit User 5000 <5000@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0